REBASEME degree()

+++ test/msp/lib/Pareto/Less.cpp
@@ -26,7 +26,7 @@ operator()(Path const* lhs,
 	weight_vector_t const& l = lhs->weight();
 	weight_vector_t const& r = rhs->weight();
 
-	const int dims = lhs->degree();
+	const size_t dims = lhs->degree();
 	assert(dims == rhs->degree());
 
 	for (int i = 0; i < dims; i++) {
